requirements.txt
# Novel Writing Assistant-Agent Pro Dependencies
# Python 3.12.x

# Core
pydantic>=2.10.6
pyyaml>=6.0
requests>=2.32.0
cryptography>=46.0.0
beautifulsoup4>=4.12.0    # 热榜爬虫 HTML 解析
lxml>=5.0.0               # BeautifulSoup 解析器后端

# LLM Client
openai>=1.60.0

# Vector Store (L2温记忆)
lancedb>=0.12.0
sentence-transformers>=2.2.2

# Scheduling (每日冥想定时任务)
apscheduler>=3.10.0

# GUI
sv-ttk>=2.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Type Checking
mypy>=1.8.0

# Development
black>=24.0.0
isort>=5.13.0
